Explicitly cast uint->int to shut up gcc warning

+++ src/gm.c
@@ -97,7 +97,7 @@ static void draw_handler(Evas_Object* choicebox __attribute__((unused)),
     char buf[256];
     struct bookinfo_t *bookinfo;
 
-    if(item_num <= MAIN_MENU_SIZE && main_menu[item_num].icon_signal)
+    if(item_num <= (int)MAIN_MENU_SIZE && main_menu[item_num].icon_signal)
         edje_object_signal_emit(item, main_menu[item_num].icon_signal, "");
     else
         edje_object_signal_emit(item, "set-icon-none", "");
